requirements.txt
fastapi[all]
pymongo
uvicorn
rich
cohere
openai
groq
python-dotenv
starlette
googlesearch-python
edge_tts
nara
mtranslate
aiofiles
python-multipart
google-generativeai
pytz
datetime
psutil
PyNaCl
pydub
flask
ffmpeg-python
requests
youtube_dl
SpeechRecognition
librosa
soundfile
numpy
pillow
opencv-python
ipython
requests
qrcode
pymongo
certifi
discord-ext-voice-recv
humanfriendly
pytube
python-dotenv
aiofiles
aiosqlite
asyncpg
asteval
rich
cryptography
httpx
